+++ training_scripts/cli.py
@@ -1,12 +1,9 @@
 """Application-wide command line utilities"""
 
-import os
-import json
-
 import click
 
 import training_scripts.application.data_generation_service as data_generation_service
-import training_scripts.application.data_balancing_service as data_balancing_service
+import training_scripts.application.dataset_generator_service as dataset_generator_service
 import training_scripts.domain.dataframe_balancing_service as dataframe_balancing_service
 
 
@@ -15,6 +12,7 @@ import training_scripts.domain.dataframe_balancing_service as dataframe_balancin
 @click.option('--sample_size', required=True, type=click.FLOAT)
 @click.option('--hours', type=click.INT, default=0)
 @click.option('--days', type=click.INT, default=0)
+@click.option('--geographically_balance_data', default=False)
 def generate_data(output_directory, sample_size, hours, days):
     """
     Function to generate and save training / testing data.
@@ -30,35 +28,10 @@ def generate_data(output_directory, sample_size, hours, days):
     generator.generate_data(output_directory, sample_size, hours, days)
 
 
-def load_files_from_directory(directory):
-    """Function to load JSONs of data from a directory."""
-    file_contents = []
-    for file_path in os.listdir(directory):
-        full_file_path = os.path.join(directory, file_path)
-        if os.path.isdir(full_file_path):
-            continue
-        with open(full_file_path, 'r') as file:
-            file_content = json.load(file)
-        file_contents.append(file_content)
-    return file_contents
-
-
 @click.command()
 @click.option('--data_directory', required=True, type=click.STRING)
-@click.option('--output_file', type=click.STRING, default='dataset.balanced')
-def generate_balanced_dataset(data_directory, output_file):
-    keys_to_balance = (
-        'GDELT',
-        'Actor1Geo_CountryCode'
-    )
-    keys_to_keep = [
-        ('GDELT', 'GlobalEventID'),
-        ('GDELT', 'Actor1Geo_CountryCode'),
-        ('GDELT', 'SOURCEURL')
-    ]
-    n_means = 5
-    data_dicts = load_files_from_directory(data_directory)
-    dataframe_balancer = dataframe_balancing_service.DataFrameBalancingService(n_means)
-    data_balancer = data_balancing_service.DataBalancingService(dataframe_balancer)
-    df = data_balancer.balance_data(data_dicts, keys_to_balance, keys_to_keep)
-    df.to_csv(output_file)
+@click.option('--output_file', 'output_path', type=click.STRING, default='dataset.balanced')
+def generate_balanced_dataset(data_directory, output_path):
+    dataframe_balancer = dataframe_balancing_service.DataFrameBalancingService(5)
+    dataset_generator = dataset_generator_service.DatasetGeneratorService(dataframe_balancer)
+    dataset_generator.balance_data(data_directory, output_path)
